batch: Move attribution diff projection
Attribution fingerprints now live outside the file attribution builder,
while attribution still owns both unit construction and displayed-diff
projection.

Selected-hunk filtering only needs to project existing attribution onto
cached diff lines, so keeping that projection in the builder module couples
display filtering to ownership construction internals.

This commit addresses that by moving projection and owned-fragment
filtering into batch.attribution_projection and pointing selected-change
filtering at the new module.

Subsequent commits will update tests and add an architecture guard for the
projection boundary.
